+++ CHANGELOG.md
@@ -5,6 +5,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.5] - 2026-09-04
+
+### Added
+- error/edge case handling
+- quit button
+- show/hide on start button
+
+## [0.2.4] - 2026-09-03
+
+### Added
+- created/more or less finished the encoder/`ffmpeg_controller`
+- Connected up the ui for the share tab and implemented it.
+
+### Fixed
+- fixed bad config.json dir
+
+
+## [0.2.3] - 2026-09-01
+
+### Added
+- created the encoder
+- added a share tab + tray menu stuff to use it
+
+### Fixed
+- removed excess files
+
 ## [0.2.2] - 2026-08-15
 
 ### Added